fix: Issue with similar function names
For the create new googleslides function, it is called create_slides. However, to create a new page in googleslides, the function was originially called create_slide. This would lead to confusion in the future. So to prevent confusion, we would call the function to create a new slide in deck create_slide_page.

See also: #8